Sample AIE-PLIO based GEMM application
The AI engine with PL example demonstrates how to use AI engine for scalar
computation, and use PL for data movement. In this example, to implement matrix
multiplication on AI engine, we use standard matrix multiplication algorithm.
The user can change the matrix size and the number of cores utilized at
compile-time. The expected matrix size must be a multiple of 50 (number of
cores used) with the minimum and maximum value as 100x100 and 800x800
respectively. Please note that this example is intended to be a proof of
concept only. There can be other ways of implementation, which can leverage
more of the AIE resource and hence can result in better performance figures.

Signed-off-by: Nishad Saraf <nishad.saraf@xilinx.com>